backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Supabase
supabase==2.0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# LangChain & RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
faiss-cpu==1.7.4

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
python-pptx==0.6.23

# Text Processing
tiktoken==0.5.2
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# File handling
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Date handling
python-dateutil==2.8.2

# Environment
environs==10.0.0

# Async support
asyncio==3.4.3
